Formatting and Codec Changes
Added some code to detect audio codec so that way there would not be a 'BitDepth' error for lossy formats and files that don't generate a bitdepth field from mediainfo. Also fixed some formatting issues with template, took off a lot of the dots (.) that separate the label from content. Also added bitrate field for audio.